삭제 확인 웹 페이지 메시지 추가

+++ BBS/src/main/webapp/view.jsp
@@ -120,7 +120,7 @@
 				if(userID != null && userID.equals(bbs.getUserID())){ 
 			%>
 				<a href="update.jsp?bbsID=<%= bbsID %>" class="btn btn-primary">수정</a>
-				<a href="deleteAction.jsp?bbsID=<%= bbsID %>" class="btn btn-primary">삭제</a>
+				<a onclick="return confirm('정말로 삭제하시겠습니까?')" href="deleteAction.jsp?bbsID=<%= bbsID %>" class="btn btn-primary">삭제</a>
 			<%
 				}
 			%>
